Always use "$@" for file name
From vanicat's pull request

The posix shell do not interpret local files = $@ as a way to set file as
an array. So when there is more than one files, or whin the file name contain a space, this will fail

It simpler to always use "$@" to manipulate the rest of the
command line, and it will correctly handle file name with space.